third_party/blink/web_tests/fast/forms/date/no-page-popup-controller-expected.txt

Confirm that the main window doesn't have pagePopupController, which should be present only in a calendar picker.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.pagePopupController === undefined is true
PASS successfullyParsed is true

TEST COMPLETE
